FFS-3007-2: Match names between report and indexing data (#807)
This new event (IncomeSummaryMatchedAgencyNames) tracks, upon delivery,
the overlap of names from the agency's original data to the names we get
from Argyle/Pinwheel.

Currently, only AZ provides us the expected name(s) of who will be using
CBV, but this is implemented generically so that we can compute this for
a future agency as well.

This creates a Service object to do the comparison, a Job class that
orchestrates fetching the report data and everything, and some tests to
make sure it works somewhat well.