Wait for browser idleness before asserting page contents
There is a race condition happening here. One theory is that the browser
is doing some navigation or something that messes with the ability to
assert the contents of the page.

To minimize the likelihood of this, let's try waiting for the browser to
be "idle". This will slow things down a bit, depending on what "idle"
means, but hopefully will reduce the chances of this race condition.