Upgrade `erzz/dockle-action` v1.3.1 -> v1.4.0
We've been having intermittent issues with the older version, possibly
because the Dockle download fails.

The newer version uses the GH Auth token when downloading Dockle, which
should be less likely to be rate-limited.

Also, I've relaxed the version constraint here to take advantage of
other upgrades like this from the community.