update translations and change the file (#859)
Added spanish translations for LA wind-down page
Fixed a couple issues with the import translation script
It assumed that you were importing a file with Windows UTF-8 encodings which caused it to break. I turned that off and added a comment.
I forced the to_yaml to write single lines instead of multi-line to reduce diffs on PR's.