FFS-3197: Reenable turbo on /employer_search by fixing style issue in CSP
We've had to special-case the /employer_search page from the rest of our
page navigations due to the Content Security Policy. Since Argyle and
Pinwheel both add `<style>` tags to the page, we need some way to ensure
those are allowed.

Previously, we just made sure the page was served without a CSP header.
This was brittle (we had to remember to disable turbo navigation on any
link that led to /employer_search so that the page would actually be
re-requested without the header) and less secure than just properly
allowing these `<style>` tags through the CSP.

The most straightforward way is to just pin the hash of the `<style>`
tag contents. This commit, primarily, does that. (This will break
whenever we update the Pinwheel/Argyle SDKs, but I've provided
instructions on how to get the new hashes at that time.) I also had to
add Argyle's domain to the script-src and connect-src.

Since that allows for re-enabling of Turbo, this commit also does that,
resulting in a much faster and smooth experience going between these
pages.