Sync from DSAC Main Sep 8 2025 (#53)
* FFS-3178: Payment Details UX changes from FFS-3027 (#908)

* Publish July 2025 Cumulative Data Update (#928)

Co-authored-by: pperozo <patricia.perozo@gmail.com>

* Fix AutoTitleTestHelper to use HTML parser instead of regex (#932)

This came up because I'm about to add an `<h1>` tag that has a class on
it, which won't match the H1_REGEX here.

* Upgrade sass from 1.89.2 to 1.90.0 (#926)

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* FFS-3219 & FFS-3329: /success shareable link updates (simplified with origin) (#930)

* FFS-3298: Return containers and database to standard configuration (#934)

* FFS-3263: replace amazon with amazon flex on popular providers (#929)

* Rename AlertComponent to Uswds::Alert and use it for site alerts (#931)

Create a new namespace to indicate this component is straight-up USWDS
markup and not something we've made for our own purposes.

* FFS-2543: Use JS to show help alert on /employer_search without refresh (#933)

* Remove the poorly-factored `check_help_param` method from
  ApplicationController.
* Remove support for the `alert_type` flash param since this was the
  only place that wasn't an "error" type.
* Render the "help" alert on the /employer_search page by default in a
  hidden state and configure Stimulus to reveal it after an aggregator
  modal is closed.
* Tweak styling on the `<h1>` as they are not expecting to have anything
  above them, and the padding is excessive when the alert is hidden.

* FFS-3209: Reconfigure database backups (#935)

* fix: upgrade @rails/actioncable from 7.2.201 to 7.2.202 (#938)

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* FFS-3197: Reenable turbo on /employer_search by fixing style issue in CSP (#936)

We've had to special-case the /employer_search page from the rest of our
page navigations due to the Content Security Policy. Since Argyle and
Pinwheel both add `<style>` tags to the page, we need some way to ensure
those are allowed.

Previously, we just made sure the page was served without a CSP header.
This was brittle (we had to remember to disable turbo navigation on any
link that led to /employer_search so that the page would actually be
re-requested without the header) and less secure than just properly
allowing these `<style>` tags through the CSP.

The most straightforward way is to just pin the hash of the `<style>`
tag contents. This commit, primarily, does that. (This will break
whenever we update the Pinwheel/Argyle SDKs, but I've provided
instructions on how to get the new hashes at that time.) I also had to
add Argyle's domain to the script-src and connect-src.

Since that allows for re-enabling of Turbo, this commit also does that,
resulting in a much faster and smooth experience going between these
pages.

* Upgrade webpack from 5.101.0 to 5.101.1 (#937)

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* FFS-N/A: Update Docker PostgreSQL (#942)

* Remove GitHub Markdown Link Check

* upgrade buttons to use https instead of http to match csp rules

* fix erblint error.

---------

Co-authored-by: Daphne Gold <daphnegold@navapbc.com>
Co-authored-by: Tom Dooner <tomdooner@navapbc.com>
Co-authored-by: pperozo <patricia.perozo@gmail.com>
Co-authored-by: decause-gov <107957201+decause-gov@users.noreply.github.com>
Co-authored-by: snyk-bot <snyk-bot@snyk.io>
Co-authored-by: Tim Miller <timmiller@navapbc.com>