Let `bin/destroy-pr-environment` succeed if there's no env to destroy (#948)
If we fail to select a workspace (used for PR apps) due to it not
existing, that's probably just because we never made a PR review app for
that branch.

In that case, let's just bail and consider it a success, since I hate
seeing red "x"s from Github Actions.