Create `cbv_flow_invitation_id` column on `cbv_applicants`
This is the first step in a data migration to improve our data model,
which is necessary for FFS-3325.

Currently, a CbvApplicant `has_many` CbvFlowInvitations, but this is not
right -- an applicant stores the data for exactly one invitation. (Maybe
the idea was that if the state sent us the same indexing data multiple
times, we would deduplicate the applicant and have it associated to many
invitations? Anyway, we're not going to do that for PII reasons.)

By moving the `cbv_flow_invitation_id` colmun to `cbv_applicants`, it
mirrors how the `cbv_flow` relates to the `cbv_applicant` and means we
can do `cbv_applicant.cbv_flow_invitation` (rather than
`cbv_applicant.cbv_flow_invitations`).

A backfill is required. This is done as a migration here so as to make
it easier for DPW to stay up-to-date with our code (compared to a
backfill script, which must be manually run inbetween deploys.)