FFS-3323: Better employment picking (#945)
Argyle sometimes sends us multiple employments. Previously, we picked the first one, which was rather naive. Now, we try to be smarter and pick a more relevant one. The complicating factor here is that we need to filter identities, incomes, and paystubs to make sure they match the employment we pick.