FFS-3419: Allow all Tokenized Links to be reused, including in LA
We had a setting, allow_invitation_reuse, which would enable a tokenized
link to be reused. We recently forgot to set this value to true for LA,
since we've updated our product thinking such that this behavior is the
default.

As such, let's simplify the logic here so that these links can always be
reused.