Allow aws / terraform commands to output to stdout / stderr (#1009)
There is currently a failed action without an error message because of
the way line 20 redirects stderror to /dev/null. Additionally the `||
true` prevents the script from aborting on the error. This change
removes the redirection so we can get a better sense of what happens
when this script fails.