Add chromedriver upgrade to `bin/update` (#1052)
This script is meant to upgrade all dependencies. Since chromedriver is
particularly nagging, let's see if we can make it automatically upgrade
using homebrew.

Not sure if the `xattr` thing will work to get around the security
thing, but would be awesome if it did. (I tried it once successfully.)