Fix product name in `<title>` tag and caseworker dashboard
We hardcode the product name ("Report My Income") in a bunch of other
places, here's a quick find-and-replace to clean up other mentions of
the old product name ("SNAP Income Pilot"). Although these Spanish
translations could be more optimal, as the caseworker dashboard is
deprecated, it didn't seem worth it to me to send this out for
re-translation.