Fix AWS region detection by using the config file
This attempts a belt-and-suspenders approach to defining the AWS region
in both an environment variable (AWS_REGION) and the config file
(~/.aws/config, as managed by `aws configure`).

This is necessary because in 29e57fc I tried to simplify region
detection script, but unfortunately failed to realize that `aws
configure get region` won't read the value from the environment
variable.