FFS-3427: Use different task definition "family" for solid_queue (#1111)
We've gotten errors from AWS when deploying:

> Too many concurrent attempts to create a new revision of the specified
> family

This is because both task definitions (for the app process and the
solid_queue workers) have the same "family" value.

By using a different value for `family`, we can avoid this concurrency
safeguard while also making it more obvious in the AWS UI which task
definitions should be assigned to which type of task.