FFS-3578: Coerce JSON API `pay_gross` to zero if null
We told LA that this field was always going to be present (in the JSON
schema we sent them). A user just had a paystub with no pay, and Argyle
returned it with a `null` value for `gross_pay_amount`.

The paystub had `earnings` entries corresponding to 12 hours worked for
"Unscheduled Time Off No Pay". Therefore, coercing to zero seems like an
accurate workaround.