Refactor `AggregatorDataHelper` logic into service object (#1149)
The `Cbv::AggregatorDataHelper` has been a thorn in our side for a
while: it has become the standard way to instantiate an AggregatorReport
class and therefore has been mixed into seven places.

That's not optimal because the AggregatorDataHelper operates via side
effects - it requires an `agency_config` and `@cbv_flow` to be defined
wherever it's mixed in, and it does its work by setting the
`@aggregator_report` instance variable.

This commit is a straight-up relocation of its logic into a service
object, to make it easier to instantiate reports in future locations.