Add `debugging:output_json_report` rake task (#1150)
We've had some requests to debug the JSON body from the JsonTransmitter.
This commit adds a rake task that can output this body.

It requried refactoring the `payload` into a separate method that could
be called indpendently of the HTTP stuff. Also, since the output will be
logged in production, I used parameter filtering so we don't
accidentally log PII.