Fix /missing_results to link to /other_jobs
Commit f31023c7 did the same thing for /synchronization_failures. Upon
further inspection, this bug is also happening to users who link a
payroll account then end up on /missing_results when trying to find a
subsequent job.

This caused an error with LA since we assumed the `has_other_jobs` field
would always have a value. To ensure the user always enters this value,
let's direct the user to /other_jobs instead.