Fix agency detection to prefer `@cbv_flow` again
I noticed a problem in local dev where the agency wasn't being set
correctly: although a "sandbox" CbvFlow was created for my flow, the
`current_agency` method was returning "az_des". This was due to a
misconfiguration of domain names in my local development environment,
but the bigger issue is that we should always be preferring the
`@cbv_flow`'s agency to the domain name.

That's the intent of the logic in the `current_agency` method, at least,
but in the flow refactor (93c23fee), we moved the `current_agency`
into `ApplicationController`. Since `ApplicationController` calls
`current_agency` in the `check_if_pilot_ended` before_action, the child
controller has not yet set `@cbv_flow`.

We should enforce the `current_agency` based on the `@cbv_flow` if
present. This commit updates the logic to do so by moving the
`check_if_pilot_ended` logic into the CBV flow base controller, so it
can run after `set_cbv_flow`.